% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_raster.R
\name{create_random_raster}
\alias{create_random_raster}
\title{Creates Random Rasters of Traces of a Specific Class.}
\source{
The function relies on the R packages \code{raster}
(\insertCite{Hijmans.2017}{TraceIdentification}) and \code{sinkr}
(\insertCite{Taylor.2017}{TraceIdentification}).
}
\usage{
create_random_raster(n.col.raster = 200, n.row.raster = 200,
  colours.cells = c("lightgrey", "green", "red"), seed = c(1:4),
  directory.result)
}
\arguments{
\item{n.col.raster}{A numeric value representing the number of columns
of the raster to create.}

\item{n.row.raster}{A numeric value representing the number of rows
of the raster to create.}

\item{colours.cells}{A character vector with colours to use for each class.}

\item{seed}{A numeric vector with integers. For each element of \code{seed},
a random raster is created.}

\item{directory.result}{A character string representing the directory where
the results will be stored.}
}
\value{
The function returns nothing, but stores a jpg file of each created
raster into \code{directory.result}. Created rasters have a resolution of
6000 times 6000 pixels and are named as raster_position_x_y where x is the
respective element of \code{seed} and y the number of elements in
\code{colours.cells}, i.e. the number of different classes.
}
\description{
\code{create_random_raster} creates a set of different rasters with a specified
number of rows and columns. For a set of classes, randomly 1 % of the cells are
selected for each class and filled with the respective colour value. 20 % of these
cells will be enlarged by the next pixel to the right, to the top and to the top right.
Each created raster will be exported as jpg file with a fixed resolution of
6000 times 6000 pixels. The funciton is used in order to create random rasters that serve
as templates in order to collect test and training samples.
}
\note{
Since the function \code{pos2coord} is part of the package \code{sinkr},
but not implemented in its structure, the R script has to be sourced prior
using \code{create_random_raster}. The function \code{pos2coord} can be retrieved
from the github repository of the author of \code{sinkr}:
\url{https://github.com/marchtaylor/sinkr/blob/master/R/pos2coord.R}.
}
\examples{
#
}
\references{
\insertAllCited{}
}
\seealso{
.
}
