% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagsvm_classification_predict.R
\name{dagsvm_classification_predict}
\alias{dagsvm_classification_predict}
\title{Classifies Images with a DAG-SVM.}
\source{
The function relies on the R packages \code{raster}
(\insertCite{Hijmans.2017}{TraceIdentification}) and \code{e1071}
(\insertCite{Meyer.2015}{TraceIdentification}).
}
\usage{
dagsvm_classification_predict(optimparam, optimmodels, directorydata)
}
\arguments{
\item{optimparam}{A \code{data.frame} object corresponding to the first element of a
list returned by \code{\link{tunedagsvm_extractoptimisedparametersandmodels}} with the
tuned parameters of the DAG-SVM.}

\item{optimmodels}{A list corresponding to the second element of a
list returned by \code{\link{tunedagsvm_extractoptimisedparametersandmodels}} with the
tuned models of the DAG-SVM.}

\item{directorydata}{A character string representing the directory where
the image files to be classified are stored. No other files
must be in this directory.}
}
\value{
The function returns a list with an element for each image file in \code{directorydata}
containing the classified image.
}
\description{
\code{dagsvm_classification_predict} takes a tuned DAG-SVM and the respective paramters
as returned by \code{\link{tunedagsvm_extractoptimisedparametersandmodels}} and a directory
name with images to classify and returns the images that were classified with the DAG-SVM.
The construction of DAG-SVM is derived according to \insertCite{Platt.2000}{TraceIdentification}
}
\examples{
#
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{tunedagsvm}}, \code{\link{tunedagsvm_extractoptimisedparametersandmodels}}.
}
