% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getclassestocompare.R
\name{getclassestocompare}
\alias{getclassestocompare}
\title{Identifies the Class Pair that is Discriminated by a Specific Node of a DAG-SVM.}
\usage{
getclassestocompare(classes, layer_i, node_i)
}
\arguments{
\item{classes}{A character vector with the class labels for each class.}

\item{layer_i}{A numeric value representing the index of the layer.}

\item{node_i}{A numeric value representing the index of the node.}
}
\value{
The function returns a character vector with the first element representing
the positive clas and the second element the negative class the SVM of \code{node_i}
of \code{layer_i} distingiushes.
}
\description{
\code{getclassestocompare} determines for a set of classes and a specific node of a
directed acyclic graph-support vector machine (DAG-SVM) (\insertCite{Platt.2000}{TraceIdentification})
the pair of classes that is
compared. This function is designed as internal function of \code{\link{tunedagsvm}}.
}
\examples{
#
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{tunedagsvm}}.
}
