% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettestandtrainingsampledf.R
\name{gettestandtrainingsampledf}
\alias{gettestandtrainingsampledf}
\title{Creates Random Rasters of Traces of a Specific Class.}
\source{
The function relies on the R packages \code{raster}
(\insertCite{Hijmans.2017}{TraceIdentification}).
}
\usage{
gettestandtrainingsampledf(directorydata, classes, subsamplesize = NULL,
  classestomerge = NULL)
}
\arguments{
\item{directorydata}{A character string representing the directory where
the image files of the test and training samples are stored. No other files
must be in this directory.}

\item{classes}{A character vector with the class labels to use for each class.
The order has to match the order of the files when assessed via
\code{\link[base]{list.files}}. The number of elements of \code{classes} has to match
the number of files in \code{directorydata}.}

\item{subsamplesize}{A numeric vector with the sizes of the subsamples to collect
for each class. The order has to match the order in  \code{classes}. The number of
elements of \code{subsamplesize} has to match the number of files in \code{directorydata}.
The order has to match the order of the files when assessed via
\code{\link[base]{list.files}}. Alternatively, \code{subsamplesize}
can be set to \code{NULL} if no subsampling should be performed.}

\item{classestomerge}{A list with a vector for each merge of two classes.
Each vector must be a character vector with two elements correspnding to two
elements of \code{classes}. The first element represents the class label to
retain whereas the class label of elements correspnding to the second element
will be changed into the first element of the vector.}
}
\value{
The function returns a \code{data.frame} object with four columns:
\describe{
  \item{\code{r}}{The values of the red band of the test and training samples.}
  \item{\code{g}}{The values of the greeb band of the test and training samples.}
  \item{\code{b}}{The values of the blue band of the test and training samples.}
  \item{\code{b}}{The corresponding class labels as factor.}
}
}
\description{
\code{gettestandtrainingsampledf} creates one \code{data.frame} object for
a set of test and training samples that can be used in combination with
\code{\link{tunedagsvm}} in order to tune a support vector machine (SVM)
in order to classify values into classes. For each class, there must
exist a an individual image file in a separate directory without any other
files. NULL values must have the rgb values \code{c(255,255,255)}, i.e.
filled with white color. It is possible to subsample the provided samples
and to merge classes.
}
\examples{
#
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{tunedagsvm}}.
}
