% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_detection.R
\name{object_detection}
\alias{object_detection}
\title{Identifies Objects/Traces in Classified Images.}
\source{
The function relies on the R packages \code{raster}
(\insertCite{Hijmans.2017}{TraceIdentification}) and \code{EBImage}
(\insertCite{Pau.2010}{TraceIdentification}).
}
\usage{
object_detection(classifiedimage, targetclassvalue = 1)
}
\arguments{
\item{classifiedimage}{A \code{\link[raster]{raster}} object (RasterLayer) representing
an image that was classified using \code{\link{dagsvm_classification_predict}}.}

\item{targetclassvalue}{A numeric value representing the class values of
\code{classifiedimage} for which objects/traces will be identified.}
}
\value{
The function returns a list with an element for each image file in \code{directorydata}
containing the classified image.
}
\description{
\code{object_detection} detects objects in classified images as clusters of pixels
having a specific value (and corresponding to a specific class) and returns their
centroid (median) coordinates. The function relies
on the funciton \code{\link[EBImage]{bwlabel}} in order to deliniate objects/traces
and on the function \code{\link[raster]{rasterToPoints}} in order to identify objects/traces
centroid coordinates.
}
\examples{
#
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{dagsvm_classification_predict}}.
}
