% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/tunedagsvm_extractoptimisedparametersandmodels.R
\name{tunedagsvm_extractoptimisedparametersandmodels}
\alias{tunedagsvm_extractoptimisedparametersandmodels}
\title{Extracts the DAG-SVM of a set of tuned DAG-SVM with the Best Classification Accuracy.}
\usage{
tunedagsvm_extractoptimisedparametersandmodels(optimparam,
  kernel = "radial")
}
\arguments{
\item{optimparam}{A list as returned by \code{\link{tunedagsvm}} containing performance
parameters and tuned DAG-SVM.}

\item{kernel}{A character value indicating a specific kernel for which to extract
the tuned SVM for each node of the DAG-SVM. If \code{kernel = NULL},
\code{tunedagsvm_extractoptimisedparametersandmodels} select for each node the kernel
with the maximum overall classification accuracy. \code{kernel} is intended to provide the
possibility to select a specific desired kernel for a DAG-SVM.}
}
\value{
The function returns a list with two elements:
\describe{
  \item{\code{prunedoptimparam}}{A \code{data.frame} object containing for each node the
  respective selected row of the first element of \code{optimparam}.}
  \item{\code{prunedbestmodels}}{A list with an element for layer containing the SVM for all
  nodes of this layer. Each element of \code{prunedbestmodels} is a list of the respective SVM.}
}
}
\description{
\code{tunedagsvm_extractoptimisedparametersandmodels} takes the output of
\code{\link{tunedagsvm}}, i.e. a set of tuned directed acyclic graph-support vector
machines (DAG-SVM) \insertCite{Platt.2000}{TraceIdentification} for a specific
classification task and returns for each node the SVM
that had the maximum overall classification accuracy when running \code{\link{tunedagsvm}}.
}
\examples{
#
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{tunedagsvm}}.
}
